service-workers/service-worker/clients-get.https.html

<!DOCTYPE html>
<title>Service Worker: Clients.get</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
const SCOPE = 'resources/clients-get-frame.html';

promise_test(async t => {
  const SCRIPT = 'resources/clients-get-worker.js';
  const client_ids = [];
  const registration =
      await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
  await wait_for_state(t, registration.installing, 'activated');
  const frame1 = await with_iframe(SCOPE + '#1');
  frame1.focus();
  let client_id = await wait_for_clientId();
  client_ids.push(client_id);
  const frame = await with_iframe(SCOPE + '#2')

  client_id = await wait_for_clientId();
  client_ids.push(client_id, 'invalid-id');
  const channel = new MessageChannel();
  frame.contentWindow.navigator.serviceWorker.controller.postMessage(
      {port: channel.port2, clientIds: client_ids}, [channel.port2]);

  const saw_message = await new Promise(resolve => {
    channel.port1.onmessage = resolve;
  });

  add_completion_callback(async() => {
    if (registration)
      await registration.unregister();
    if (frame1)
      frame1.remove();
    if (frame)
      frame.remove();
  });

  assert_equals(saw_message.data.length, 3);
  assert_array_equals(saw_message.data[0], expected[0]);
  assert_array_equals(saw_message.data[1], expected[1]);
  assert_equals(saw_message.data[2], expected[2]);
});

function wait_for_clientId() {
  return new Promise(function(resolve, reject) {
      function get_client_id(e) {
        window.removeEventListener('message', get_client_id);
        resolve(e.data.clientId);
      }
      window.addEventListener('message', get_client_id, false);
    });
}

const expected = [
    // visibilityState, focused, url, type, frameType
    ['visible', true, normalizeURL(SCOPE) + '#1', 'window', 'nested'],
    ['visible', false, normalizeURL(SCOPE) + '#2', 'window', 'nested'],
    undefined
];
</script>
